DT-1788 Improve reporting (#25)
* DT-1788 WIP Full Sync process

* DT-1788 WIP Full Sync process

* DT-1788 WIP Full Sync process

* DT-1788 WIP Full Sync process

* DT-1788 WIP Full Sync process

* DT-1788 Upper case description

* DT-1788 Sync service

* DT-1788 Sync service

* DT-1788 Sync service

* DT-1788 Sync service reporting improvements

* DT-1788 Klint

* DT-1788 Klint

* DT-1788 Handle skip location description

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsregisterstonomisupdate/services/CourtRegisterUpdateService.kt
@@ -1,18 +1,25 @@
 package uk.gov.justice.digital.hmpps.hmppsregisterstonomisupdate.services
 
-import com.fasterxml.jackson.annotation.JsonIgnore
+import com.fasterxml.jackson.annotation.JsonInclude
+import com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL
 import com.google.common.collect.MapDifference
 import com.google.common.collect.Maps
 import com.google.common.reflect.TypeToken
 import com.google.gson.Gson
 import com.microsoft.applicationinsights.TelemetryClient
+import io.swagger.v3.oas.annotations.media.Schema
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
 import uk.gov.justice.digital.hmpps.hmppsregisterstonomisupdate.model.CourtUpdate
+import uk.gov.justice.digital.hmpps.hmppsregisterstonomisupdate.services.CourtDifferences.UpdateType.ERROR
+import uk.gov.justice.digital.hmpps.hmppsregisterstonomisupdate.services.CourtDifferences.UpdateType.INSERT
+import uk.gov.justice.digital.hmpps.hmppsregisterstonomisupdate.services.CourtDifferences.UpdateType.NONE
+import uk.gov.justice.digital.hmpps.hmppsregisterstonomisupdate.services.CourtDifferences.UpdateType.UPDATE
 import java.lang.reflect.Type
 import java.time.LocalDate
+import java.util.SortedSet
 
 @Service
 class CourtRegisterUpdateService(
@@ -39,8 +46,8 @@ class CourtRegisterUpdateService(
     "OTH" to "OTHER"
   )
 
-  fun updateCourtDetails(court: CourtUpdate): UpdateStatistics {
-    val stats = UpdateStatistics()
+  fun updateCourtDetails(court: CourtUpdate): SyncStatistics {
+    val stats = SyncStatistics()
 
     courtRegisterService.getCourtInfoFromRegister(court.courtId)?.run {
 
@@ -49,6 +56,7 @@ class CourtRegisterUpdateService(
         processCourt(it, stats)
       }
     }
+    log.debug("Sync Stats: $stats")
     return stats
   }
 
@@ -81,7 +89,7 @@ class CourtRegisterUpdateService(
   fun String.truncate(maxChar: Int): String =
     if (this.length < maxChar) this else this.substring(0, maxChar)
 
-  private fun processCourt(courtDto: CourtDataToSync, stats: UpdateStatistics) {
+  private fun processCourt(courtDto: CourtDataToSync, stats: SyncStatistics) {
 
     val currentCourtDataInPrisonSystem = prisonService.getCourtInformation(courtDto.courtId)
 
@@ -95,32 +103,30 @@ class CourtRegisterUpdateService(
   fun syncCourt(
     currentCourtDataToCompare: CourtDataToSync?,
     newCourtData: CourtDataToSync,
-    stats: UpdateStatistics
+    stats: SyncStatistics
   ) {
 
     mergeIds(newCourtData, currentCourtDataToCompare)
 
     val diff = checkForDifferences(currentCourtDataToCompare, newCourtData)
-    stats.diffs.add(diff)
+
     if (!diff.areEqual()) {
+      stats.courts[newCourtData.courtId] = CourtDifferences(newCourtData.courtId, diff.toString())
       try {
-        stats.add(storeInPrisonData(currentCourtDataToCompare, newCourtData, applyChanges))
-        val trackingAttributes = mapOf(
-          "courtId" to newCourtData.courtId,
-          "changes" to diff.toString(),
-          "changes-applied" to applyChanges.toString()
-        )
-        telemetryClient.trackEvent("HR2NU-Court-Change", trackingAttributes, null)
+        storeInPrisonData(currentCourtDataToCompare, newCourtData, stats)
+        if (applyChanges && stats.courts[newCourtData.courtId]?.updateType != NONE) {
+          val trackingAttributes = mapOf(
+            "courtId" to newCourtData.courtId,
+            "differences" to stats.courts[newCourtData.courtId]?.differences,
+          )
+          telemetryClient.trackEvent("HR2NU-Court-Change", trackingAttributes, null)
+        }
       } catch (e: Exception) {
-        stats.courtsFailures.add(newCourtData.courtId)
+        stats.courts[newCourtData.courtId] = stats.courts[newCourtData.courtId]!!.copy(updateType = ERROR)
+
         log.error("Failed to update {} - message = {}", newCourtData.courtId, e.message)
-        val trackingAttributes = mapOf(
-          "courtId" to newCourtData.courtId
-        )
-        telemetryClient.trackEvent("HR2NU-Court-Change-Failure", trackingAttributes, null)
+        telemetryClient.trackEvent("HR2NU-Court-Change-Failure", mapOf("courtId" to newCourtData.courtId), null)
       }
-    } else {
-      telemetryClient.trackEvent("HR2NU-Court-No-Change", mapOf("courtId" to newCourtData.courtId), null)
     }
   }
 
@@ -150,37 +156,35 @@ class CourtRegisterUpdateService(
           comment = address.comment,
           phones = address.phones.map { phone ->
             PhoneFromPrisonSystem(phone.phoneId, phone.number, phone.type, phone.ext)
-          }
+          }.toSortedSet(naturalOrder())
         )
-      }
+      }.toSortedSet(naturalOrder())
     )
 
   private fun storeInPrisonData(
     currentCourtData: CourtDataToSync?,
     newCourtData: CourtDataToSync,
-    applyChanges: Boolean = false
-  ): UpdateStatistics {
+    stats: SyncStatistics
+  ) {
 
-    val stats = UpdateStatistics()
-    if (applyChanges) {
-      val dataPayload = translateToPrisonSystemFormat(newCourtData)
-      if (currentCourtData == null) {
-        prisonService.insertCourt(dataPayload)
-        stats.courtsInserted.add(dataPayload.agencyId)
-      } else {
-        if (newCourtData != currentCourtData) { // don't update if equal
-          prisonService.updateCourt(dataPayload)
-          stats.courtsUpdated.add(dataPayload.agencyId)
-        }
+    val dataPayload = translateToPrisonSystemFormat(newCourtData)
+    if (currentCourtData == null) {
+      if (applyChanges) prisonService.insertCourt(dataPayload)
+      stats.courts[newCourtData.courtId] = stats.courts[newCourtData.courtId]!!.copy(updateType = INSERT)
+    } else {
+      if (newCourtData != currentCourtData) { // don't update if equal
+        if (applyChanges) prisonService.updateCourt(dataPayload)
+        stats.courts[newCourtData.courtId] = stats.courts[newCourtData.courtId]!!.copy(updateType = UPDATE)
       }
     }
 
     currentCourtData?.addresses?.forEach {
       if (newCourtData.addresses.find { a -> a.addressId == it.addressId } == null) {
-        if (applyChanges) {
-          prisonService.removeAddress(newCourtData.courtId, it.addressId!!)
-          stats.numberAddressesRemoved = stats.numberAddressesRemoved + 1
-        }
+        if (applyChanges) prisonService.removeAddress(newCourtData.courtId, it.addressId!!)
+        stats.courts[newCourtData.courtId] = stats.courts[newCourtData.courtId]!!.copy(
+          updateType = UPDATE,
+          numberAddressesRemoved = stats.courts[newCourtData.courtId]?.numberAddressesRemoved?.plus(1) ?: 0
+        )
       }
     }
 
@@ -192,41 +196,55 @@ class CourtRegisterUpdateService(
       currentAddress?.run {
         this.phones.forEach {
           if (updatedAddress.phones.find { p -> p.phoneId == it.phoneId } == null) {
-            if (applyChanges) {
-              prisonService.removePhone(newCourtData.courtId, this.addressId!!, it.phoneId!!)
-              stats.numberPhonesRemoved = stats.numberPhonesRemoved + 1
-            }
+
+            if (applyChanges) prisonService.removePhone(newCourtData.courtId, this.addressId!!, it.phoneId!!)
+            stats.courts[newCourtData.courtId] = stats.courts[newCourtData.courtId]!!.copy(
+              updateType = UPDATE,
+              numberPhonesRemoved = stats.courts[newCourtData.courtId]?.numberPhonesRemoved?.plus(1) ?: 0
+            )
           }
         }
       }
 
-      val updatedAddressId =
-        if (applyChanges) updateAddress(newCourtData.courtId, updatedAddress, currentAddress, stats) else null
+      val updatedAddressId = updateAddress(newCourtData.courtId, updatedAddress, currentAddress, stats, applyChanges)
 
       // update phones
       updatedAddress.phones.forEach { phone ->
         val currentPhone = currentAddress?.phones?.find { it.phoneId == phone.phoneId }
         if (phone.phoneId == null) {
-          if (applyChanges && updatedAddressId != null) {
-            prisonService.insertPhone(newCourtData.courtId, updatedAddressId, phone)
-            stats.numberPhonesInserted = stats.numberPhonesInserted + 1
+          if (updatedAddressId != null) {
+            if (applyChanges) prisonService.insertPhone(newCourtData.courtId, updatedAddressId, phone)
+            stats.courts[newCourtData.courtId] = stats.courts[newCourtData.courtId]!!.copy(
+              updateType = UPDATE,
+              numberPhonesInserted = stats.courts[newCourtData.courtId]?.numberPhonesInserted?.plus(1) ?: 0
+            )
           }
         } else {
           if (phone != currentPhone) {
-            if (applyChanges && updatedAddressId != null) {
-              prisonService.updatePhone(newCourtData.courtId, updatedAddressId, phone)
-              stats.numberPhonesUpdated = stats.numberPhonesUpdated + 1
+            if (updatedAddressId != null) {
+              if (applyChanges) prisonService.updatePhone(newCourtData.courtId, updatedAddressId, phone)
+              stats.courts[newCourtData.courtId] = stats.courts[newCourtData.courtId]!!.copy(
+                updateType = UPDATE,
+                numberPhonesUpdated = stats.courts[newCourtData.courtId]?.numberPhonesUpdated?.plus(1) ?: 0
+              )
             }
           }
         }
       }
     }
-    return stats
+
+    if (stats.courts[newCourtData.courtId]?.updateType == NONE) {
+      stats.courts.remove(newCourtData.courtId)
+    }
   }
 
-  private fun checkForDifferences(existingRecord: CourtDataToSync?, newRecord: CourtDataToSync): MapDifference<String, Any> {
+  private fun checkForDifferences(
+    existingRecord: CourtDataToSync?,
+    newRecord: CourtDataToSync
+  ): MapDifference<String, Any> {
     val type: Type = object : TypeToken<Map<String, Any>>() {}.type
-    val leftMap: Map<String, Any> = if (existingRecord != null) gson.fromJson(gson.toJson(existingRecord), type) else mapOf()
+    val leftMap: Map<String, Any> =
+      if (existingRecord != null) gson.fromJson(gson.toJson(existingRecord), type) else mapOf()
     val rightMap: Map<String, Any> = gson.fromJson(gson.toJson(newRecord), type)
     return Maps.difference(leftMap, rightMap)
   }
@@ -235,19 +253,26 @@ class CourtRegisterUpdateService(
     courtId: String,
     updatedAddress: AddressDataToSync,
     currentAddress: AddressDataToSync?,
-    stats: UpdateStatistics
+    stats: SyncStatistics,
+    applyChanges: Boolean
   ): Long? {
     val dataPayload = translateToPrisonSystemFormat(updatedAddress)
     if (dataPayload.addressId == null) {
-      val addressId = prisonService.insertAddress(courtId, dataPayload).addressId
-      stats.numberAddressesInserted = stats.numberAddressesInserted + 1
+      val addressId = if (applyChanges) prisonService.insertAddress(courtId, dataPayload).addressId else null
+      stats.courts[courtId] = stats.courts[courtId]!!.copy(
+        updateType = UPDATE,
+        numberAddressesInserted = stats.courts[courtId]?.numberAddressesInserted?.plus(1) ?: 0
+      )
       return addressId
     }
 
     currentAddress.run {
       if (this != updatedAddress) {
-        val addressId = prisonService.updateAddress(courtId, dataPayload).addressId
-        stats.numberAddressesUpdated = stats.numberAddressesUpdated + 1
+        val addressId = if (applyChanges) prisonService.updateAddress(courtId, dataPayload).addressId else null
+        stats.courts[courtId] = stats.courts[courtId]!!.copy(
+          updateType = UPDATE,
+          numberAddressesUpdated = stats.courts[courtId]?.numberAddressesUpdated?.plus(1) ?: 0
+        )
         return addressId
       }
     }
@@ -284,7 +309,7 @@ class CourtRegisterUpdateService(
       startDate = it.startDate,
       endDate = it.endDate,
       comment = it.comment,
-      phones = it.phones
+      phones = it.phones.toList()
     )
 
   private fun translateToPrisonSystemFormat(addressData: AddressDataToSync) =
@@ -304,10 +329,10 @@ class CourtRegisterUpdateService(
     }
 
     // search for matching addresses with just one address
-    if (legacyCourt.addresses.size == 1 && updatedCourtData.addresses.size == 1 && updatedCourtData.addresses[0].addressId == null) {
-      val primaryAddress = updatedCourtData.addresses[0]
+    if (legacyCourt.addresses.size == 1 && updatedCourtData.addresses.size == 1 && updatedCourtData.addresses.first().addressId == null) {
+      val primaryAddress = updatedCourtData.addresses.first()
 
-      with(legacyCourt.addresses[0]) {
+      with(legacyCourt.addresses.first()) {
         updateAddressAndPhone(primaryAddress)
       }
     }
@@ -338,37 +363,35 @@ class CourtRegisterUpdateService(
           comment = null,
           phones = building.contacts.map { phone ->
             PhoneFromPrisonSystem(null, phone.detail, if (phone.type == "TEL") "BUS" else phone.type, null)
-          }
+          }.toSortedSet(naturalOrder())
         )
-      }
+      }.toSortedSet(naturalOrder())
     )
 }
 
-data class UpdateStatistics(
-  var courtsInserted: MutableList<String> = mutableListOf(),
-  var courtsUpdated: MutableList<String> = mutableListOf(),
-  var courtsFailures: MutableList<String> = mutableListOf(),
-  var numberAddressesInserted: Int = 0,
-  var numberAddressesUpdated: Int = 0,
-  var numberAddressesRemoved: Int = 0,
-  var numberPhonesInserted: Int = 0,
-  var numberPhonesUpdated: Int = 0,
-  var numberPhonesRemoved: Int = 0,
-  @JsonIgnore
-  var diffs: MutableList<MapDifference<String, Any>> = mutableListOf()
+@JsonInclude(NON_NULL)
+@Schema(description = "Sync Statistics")
+data class SyncStatistics(
+  @Schema(description = "Map of all courts have have been inserted, updated or errored")
+  val courts: MutableMap<String, CourtDifferences> = mutableMapOf()
+)
+
+@JsonInclude(NON_NULL)
+@Schema(description = "Court Changes")
+data class CourtDifferences(
+  @Schema(description = "The ID of the Court", example = "SHFFCC") val courtId: String,
+  @Schema(description = "Differences listed", example = "SHFFCC") val differences: String,
+  @Schema(description = "Type of update", example = "INSERT") val updateType: UpdateType = NONE,
+  @Schema(description = "Number of addresses inserted for this court", example = "1") val numberAddressesInserted: Int = 0,
+  @Schema(description = "Number of addresses updated for this court", example = "2") val numberAddressesUpdated: Int = 0,
+  @Schema(description = "Number of addresses removed for this court", example = "3") val numberAddressesRemoved: Int = 0,
+  @Schema(description = "Number of phones inserted for this court", example = "1") val numberPhonesInserted: Int = 0,
+  @Schema(description = "Number of phones updated for this court", example = "2") val numberPhonesUpdated: Int = 0,
+  @Schema(description = "Number of phones removed for this court", example = "1") val numberPhonesRemoved: Int = 0
 ) {
 
-  fun add(toAdd: UpdateStatistics) {
-    courtsInserted.addAll(toAdd.courtsInserted)
-    courtsUpdated.addAll(toAdd.courtsUpdated)
-    courtsFailures.addAll(toAdd.courtsFailures)
-    numberAddressesInserted += toAdd.numberAddressesInserted
-    numberAddressesUpdated += toAdd.numberAddressesUpdated
-    numberAddressesRemoved += toAdd.numberAddressesRemoved
-    numberPhonesInserted += toAdd.numberPhonesInserted
-    numberPhonesUpdated += toAdd.numberPhonesUpdated
-    numberPhonesRemoved += toAdd.numberPhonesRemoved
-    diffs.addAll(toAdd.diffs)
+  enum class UpdateType {
+    NONE, INSERT, UPDATE, ERROR
   }
 }
 
@@ -379,7 +402,7 @@ data class CourtDataToSync(
   val active: Boolean,
   val courtType: String,
   val deactivationDate: LocalDate? = null,
-  val addresses: List<AddressDataToSync> = listOf(),
+  val addresses: SortedSet<AddressDataToSync> = sortedSetOf(),
 ) {
 
   override fun hashCode(): Int {
@@ -424,9 +447,9 @@ data class AddressDataToSync(
   val noFixedAddress: Boolean,
   var startDate: LocalDate? = null,
   var endDate: LocalDate? = null,
-  val phones: List<PhoneFromPrisonSystem> = listOf(),
+  val phones: SortedSet<PhoneFromPrisonSystem> = sortedSetOf(),
   var comment: String? = null
-) {
+) : Comparable<AddressDataToSync> {
 
   fun updateAddressAndPhone(
     address: AddressDataToSync
@@ -439,6 +462,18 @@ data class AddressDataToSync(
 
     // update the phones
     updatePhoneIds(address, this)
+  }
+
+  override fun compareTo(other: AddressDataToSync): Int {
+    return compareBy<AddressDataToSync>(
+      { it.premise },
+      { it.postalCode },
+      { it.street },
+      { it.locality },
+      { it.town?.description },
+      { it.county?.description },
+      { it.country?.description }
+    ).compare(this, other)
   }
 
   private fun updatePhoneIds(
@@ -460,12 +495,12 @@ data class AddressDataToSync(
     other as AddressDataToSync
 
     if (premise != other.premise) return false
+    if (postalCode != other.postalCode) return false
     if (street != other.street) return false
     if (locality != other.locality) return false
     if (town != other.town) return false
     if (county != other.county) return false
     if (country != other.country) return false
-    if (postalCode != other.postalCode) return false
     if (primary != other.primary) return false
 
     return true
@@ -473,12 +508,12 @@ data class AddressDataToSync(
 
   override fun hashCode(): Int {
     var result = premise?.hashCode() ?: 0
+    result = 31 * result + (postalCode?.hashCode() ?: 0)
     result = 31 * result + (street?.hashCode() ?: 0)
     result = 31 * result + (locality?.hashCode() ?: 0)
     result = 31 * result + (town?.hashCode() ?: 0)
     result = 31 * result + (county?.hashCode() ?: 0)
     result = 31 * result + (country?.hashCode() ?: 0)
-    result = 31 * result + (postalCode?.hashCode() ?: 0)
     result = 31 * result + primary.hashCode()
     return result
   }
